feat: implement issues #1069-#1072 - SDK cleanup, audit scripts, type consolidation, debug removal
- #1072: Organize SDK exports with clear comments and structure
- #1071: Add dependency audit scripts across all packages (pnpm audit, cargo audit)
- #1070: Create shared @bluecollar/types package, consolidate duplicates across app/sdk/mobile
- #1069: Replace console.log with proper logger in production code, add ESLint rules

All packages now use shared types from @bluecollar/types.
Added audit scripts: npm run audit, security:scan targets.
Production console statements replaced with structured logging.
ESLint rules prevent future console usage (except tests/seeds/scripts).
