+++ package.json
@@ -12,6 +12,9 @@
     "test": "pnpm --filter @bluecollar/api test && pnpm --filter @bluecollar/app test",
     "lint": "pnpm --filter @bluecollar/api lint && pnpm --filter @bluecollar/app lint",
     "format": "prettier --write \"packages/**/*.{ts,tsx,js,json,md}\"",
+    "audit": "pnpm audit --audit-level=high",
+    "audit:fix": "pnpm audit --fix",
+    "security:scan": "pnpm audit --audit-level=high && pnpm --filter @bluecollar/contracts security:scan",
     "docker:up": "docker compose up -d",
     "docker:down": "docker compose down"
   },
