+++ packages/api/.eslintrc.json
@@ -18,7 +18,22 @@
     "@typescript-eslint/no-unused-vars": ["warn", { "argsIgnorePattern": "^_" }],
     "@typescript-eslint/no-explicit-any": "warn",
     "@typescript-eslint/explicit-function-return-type": "off",
-    "no-console": "off"
+    "no-console": ["error", { "allow": [] }],
+    "no-debugger": "error"
   },
+  "overrides": [
+    {
+      "files": [
+        "src/database/seed*.ts",
+        "src/commands/*.ts",
+        "src/scripts/*.ts",
+        "src/monitoring/*.ts",
+        "src/__tests__/**/*.ts"
+      ],
+      "rules": {
+        "no-console": "off"
+      }
+    }
+  ],
   "ignorePatterns": ["dist/", "node_modules/"]
 }
