+++ packages/app/.eslintrc.json
@@ -1,6 +1,21 @@
 {
   "extends": ["next/core-web-vitals", "next/typescript"],
   "rules": {
-    "@typescript-eslint/no-unused-vars": ["error", { "argsIgnorePattern": "^_", "varsIgnorePattern": "^_" }]
-  }
+    "@typescript-eslint/no-unused-vars": ["error", { "argsIgnorePattern": "^_", "varsIgnorePattern": "^_" }],
+    "no-console": ["error", { "allow": ["warn", "error"] }],
+    "no-debugger": "error"
+  },
+  "overrides": [
+    {
+      "files": [
+        "src/**/*.stories.tsx",
+        "src/**/*.test.ts",
+        "src/**/*.test.tsx",
+        "public/sw.js"
+      ],
+      "rules": {
+        "no-console": "off"
+      }
+    }
+  ]
 }
