+++ packages/app/package.json
@@ -13,10 +13,13 @@
     "test:e2e:ui": "playwright test --ui",
     "lint": "next lint",
     "type-check": "tsc --noEmit",
+    "audit": "pnpm audit --audit-level=high",
+    "audit:fix": "pnpm audit --fix",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build"
   },
   "dependencies": {
+    "@bluecollar/types": "workspace:*",
     "@hookform/resolvers": "^3.9.0",
     "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-dropdown-menu": "^2.1.16",
