+++ packages/app/src/components/OnboardingModal.stories.tsx
@@ -13,15 +13,15 @@ type Story = StoryObj<typeof OnboardingModal>
 export const Default: Story = {
   args: {
     isOpen: true,
-    onComplete: () => console.log('Onboarding completed'),
-    onSkip: () => console.log('Onboarding skipped'),
+    onComplete: () => {},
+    onSkip: () => {},
   },
 }
 
 export const Closed: Story = {
   args: {
     isOpen: false,
-    onComplete: () => console.log('Onboarding completed'),
-    onSkip: () => console.log('Onboarding skipped'),
+    onComplete: () => {},
+    onSkip: () => {},
   },
 }
