+++ packages/sdk/package.json
@@ -11,6 +11,7 @@
     "test": "vitest run"
   },
   "dependencies": {
+    "@bluecollar/types": "workspace:*",
     "@stellar/stellar-sdk": "^13.0.0"
   },
   "devDependencies": {
